Bump apis to v0.0.1
