+++ cmd/release/main.go
@@ -20,6 +20,10 @@ limitations under the License.
 // prefix is the component name (for go-gettable modules it is also the module's
 // directory path, so the tag doubles as the Go-module tag):
 //
+//	golang-commons     golang-commons/v<X.Y.Z>     go-gettable module tag for
+//	                                               go.platform-mesh.io/golang-commons (no image)
+//	subroutines        subroutines/v<X.Y.Z>        go-gettable module tag for
+//	                                               go.platform-mesh.io/subroutines (no image)
 //	apis               apis/v<X.Y.Z>               go-gettable module tag for
 //	                                               go.platform-mesh.io/apis (no image)
 //	golang-commons     golang-commons/v<X.Y.Z>     golang-commons.yml: go-gettable module
@@ -90,54 +94,13 @@ import (
 	"strings"
 )
 
-// component describes a release line: its tag prefix (the version is appended
-// directly, e.g. prefix "apis/v" + "0.0.1" = "apis/v0.0.1") and a one-line
-// summary of what cutting the tag sets in motion — shown in the plan so a
-// dry-run makes the downstream effect obvious. Order in componentOrder matters
-// for `all`.
-type component struct {
-	prefix   string
-	triggers string
-}
+// The release component registry (component, componentOrder, components,
+// libraryComponents) lives in const.go.
 
-// apis is first: the operators depend on the apis module, so when releasing
-// `all` the apis tag is cut before the operators that will eventually `require`
-// that published version.
-var componentOrder = []string{
-	"apis",
-	"golang-commons",
-	"subroutines",
-	"account-operator",
-	"backup-operator",
-	"extension-manager-operator",
-	"iam-service",
-	"kcp-migration-operator",
-	"resource-sharding-operator",
-	"search-operator",
-	"search-service",
-	"security-operator",
-	"terminal-controller-manager",
-	"rebac-authz-webhook",
-	"virtual-workspaces",
-}
-
-var components = map[string]component{
-	"apis":                        {"apis/v", "go-gettable module tag for go.platform-mesh.io/apis (no image)"},
-	"golang-commons":              {"golang-commons/v", "golang-commons.yml: go-gettable module tag for go.platform-mesh.io/golang-commons, cuts a GitHub release (no image)"},
-	"subroutines":                 {"subroutines/v", "subroutines.yml: go-gettable module tag for go.platform-mesh.io/subroutines, cuts a GitHub release (no image)"},
-	"account-operator":            {"account-operator/v", "account-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"backup-operator":             {"backup-operator/v", "backup-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"extension-manager-operator":  {"extension-manager-operator/v", "extension-manager-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"iam-service":                 {"iam-service/v", "iam-service.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"kcp-migration-operator":      {"kcp-migration-operator/v", "kcp-migration-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"resource-sharding-operator":  {"resource-sharding-operator/v", "resource-sharding-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"search-operator":             {"search-operator/v", "search-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"search-service":              {"search-service/v", "search-service.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"security-operator":           {"security-operator/v", "security-operator.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"terminal-controller-manager": {"terminal-controller-manager/v", "terminal-controller-manager.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"rebac-authz-webhook":         {"rebac-authz-webhook/v", "rebac-authz-webhook.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-	"virtual-workspaces":          {"virtual-workspaces/v", "virtual-workspaces.yml: builds + signs the image, cuts a GitHub release, bumps the chart, publishes SBOM + signed OCM component"},
-}
+// plan is one component's resolved release step: the tag it moves from, the
+// full tag to create, the bare version (e.g. v0.0.1), and what cutting it
+// triggers downstream.
+type plan struct{ name, from, fullTag, ver, triggers string }
 
 func main() {
 	if err := run(os.Args[1:]); err != nil {
@@ -217,7 +180,6 @@ func run(args []string) error {
 	branch, _ := gitOut("rev-parse", "--abbrev-ref", "HEAD")
 
 	// Build the plan.
-	type plan struct{ name, from, fullTag, triggers string }
 	var plans []plan
 	for _, name := range names {
 		comp := components[name]
@@ -242,7 +204,7 @@ func run(args []string) error {
 		if hasLatest {
 			from = comp.prefix + strings.TrimPrefix(latest.String(), "v")
 		}
-		plans = append(plans, plan{name, from, full, comp.triggers})
+		plans = append(plans, plan{name, from, full, next.String(), comp.triggers})
 	}
 
 	// Show the plan: the version step and what each tag sets in motion.
@@ -258,6 +220,7 @@ func run(args []string) error {
 		for _, p := range plans {
 			fmt.Printf("  git tag %s %s && git push origin %s\n", p.fullTag, opts.ref, p.fullTag)
 		}
+		printBumpHints(plans)
 		return nil
 	}
 
@@ -286,7 +249,28 @@ func run(args []string) error {
 		fmt.Printf("pushed %s\n", p.fullTag)
 	}
 	fmt.Println("\nDone — the release workflow will pick these up.")
+	printBumpHints(plans)
 	return nil
+}
+
+// printBumpHints tells the user how to point the repo's consumers at the
+// freshly tagged library modules. cmd/release only moves the tag; bumping the
+// dependents is a separate, reviewable commit — so we print the exact
+// `task bump-deps` command rather than editing go.mod files here.
+func printBumpHints(plans []plan) {
+	var hints []string
+	for _, p := range plans {
+		if libraryComponents[p.name] {
+			hints = append(hints, fmt.Sprintf("  task bump-deps -- %s %s", p.name, p.ver))
+		}
+	}
+	if len(hints) == 0 {
+		return
+	}
+	fmt.Println("\nNow point the repo's consumers at the released version(s):")
+	for _, h := range hints {
+		fmt.Println(h)
+	}
 }
 
 // latestTag returns the highest semver tag carrying prefix, with the prefix
@@ -458,6 +442,8 @@ Usage:
   release <component|all> [flags]
 
 Components:
+  golang-commons               golang-commons/v<X.Y.Z>               (go-gettable module tag, no image)
+  subroutines                  subroutines/v<X.Y.Z>                  (go-gettable module tag, no image)
   apis                         apis/v<X.Y.Z>                         (go-gettable module tag, no image)
   golang-commons               golang-commons/v<X.Y.Z>               (go-gettable module tag + GitHub release, no image)
   subroutines                  subroutines/v<X.Y.Z>                  (go-gettable module tag + GitHub release, no image)
