+++ operators/terminal-controller-manager/go.mod
@@ -9,7 +9,7 @@ require (
 	github.com/spf13/cobra v1.10.2
 	github.com/spf13/pflag v1.0.10
 	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.69.0
-	go.platform-mesh.io/apis v0.0.0-20260622214312-00032b10b96f
+	go.platform-mesh.io/apis v0.0.1
 	go.platform-mesh.io/golang-commons v0.0.0-20260622214312-00032b10b96f
 	go.platform-mesh.io/subroutines v0.0.0-20260622214312-00032b10b96f
 	k8s.io/api v0.35.4
@@ -81,7 +81,6 @@ require (
 	go.opentelemetry.io/otel/sdk v1.44.0 // indirect
 	go.opentelemetry.io/otel/trace v1.44.0 // indirect
 	go.opentelemetry.io/proto/otlp v1.10.0 // indirect
-	go.uber.org/goleak v1.3.1-0.20251210191316-2b7fd8a0d244 // indirect
 	go.yaml.in/yaml/v2 v2.4.4 // indirect
 	go.yaml.in/yaml/v3 v3.0.4 // indirect
 	golang.org/x/net v0.55.0 // indirect
