+++ services/iam-service/go.mod
@@ -20,7 +20,7 @@ require (
 	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.69.0
 	go.opentelemetry.io/otel v1.44.0
 	go.platform-mesh.io/account-operator v0.0.0-20260622214312-00032b10b96f
-	go.platform-mesh.io/apis v0.0.0-20260622214312-00032b10b96f
+	go.platform-mesh.io/apis v0.0.1
 	go.platform-mesh.io/golang-commons v0.0.0-20260622214312-00032b10b96f
 	golang.org/x/oauth2 v0.36.0
 	golang.org/x/sync v0.21.0
@@ -30,8 +30,6 @@ require (
 	sigs.k8s.io/controller-runtime v0.23.3
 	sigs.k8s.io/multicluster-runtime v0.23.3
 )
-
-require go.uber.org/goleak v1.3.1-0.20251210191316-2b7fd8a0d244 // indirect
 
 replace github.com/kcp-dev/multicluster-provider/client => github.com/kcp-dev/multicluster-provider/client v0.7.1
